fix bug: shows connected but data was not sent to emulator see https://groups.google.com/forum/#!topic/android-developers/s6pQibI0Jb8
The problem was (and presumably still is) that a string in the protocol being sent from the emulator to the app was being corrupted with four leading zero bytes.  I didn't investigate the emulator code, but my assumption is that a struct-padding issue somewhere means that an extra word is being inserted into the stream of data when the emulator is compiled for MS Windows, but not for the Mac. (So the proper solution would be to track down where this is happening and process the relevant fields individually, rather than take the containing struct as a whole.)

But for my purposes, I came up with a hack in the app file Connection.java.  The string being corrupted is the channel name, causing the method getChannel() to return null -